Backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
django-filter==23.5 